Chore: Rename test_amd64 package
